Replace date picker
The commit replaces the previous date picker with a new one that uses MatDatePicker. It also updates the PlanningReiterationModel to accept Date objects instead of strings. Additionally, it fixes some formatting issues in the HTML and TypeScript files related to translations and form fields. Finally, it updates the updateStartDate and updateEndDate functions to use MatDatepickerInputEvent instead of any type for better type safety.