Fix Number not saving and not display in edit view
The commit fixes an issue where the number was not being saved and displayed correctly in the edit view. The changes include updating the property name from "itemNumber" to "planningNumber" in both the PlanningUpdateModel and PlanningModel classes. Additionally, the ngModel binding in the planning-edit.component.html file was also updated to use "planningNumber". Finally, in the PlanningService.cs file, the "PlanningNumber" property is now assigned with the value from the model.